+++ packages/types/custom/core.ts
@@ -425,7 +425,7 @@ export interface IconMenuConfiguration extends PluginOptions {
   initiallyOpen?: string
 }
 
-export interface LayerChooserConfiguration {
+export interface LayerChooserConfiguration extends PluginOptions {
   component?: VueConstructor
 }
 
@@ -670,8 +670,10 @@ export interface MapConfig extends MasterportalApiConfig {
   checkServiceAvailability?: boolean
   extendedMasterportalapiMarkers?: ExtendedMasterportalapiMarkers
   featureStyles?: string
+  secureServiceUrlRegex?: string
   language?: InitialLanguage
   locales?: Locale[]
+  oidcToken?: string
   renderFaToLightDom?: boolean
   stylePath?: string
   vuetify?: UserVuetifyPreset
@@ -746,6 +748,7 @@ export interface CoreState {
   mapHasDimensions: boolean
   moveHandle: number
   moveHandleActionButton: number
+  oidcToken: string
   // NOTE truly any since external plugins may bring whatever; unknown will lead to further errors
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   plugin: Record<string, any>