+++ packages/core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Add `singleTile` as as usable parameter in the configuration of WMS-layers.
 - Feature: Additionally export `MapInstance` type.
+- Feature: Add possibility of using OIDC secured services by using the configuration parameters `secureServiceUrlRegex` and `oidcToken` as well as the mutation `setOidcToken`.
 - Fix: If a flag `_isPolarDragLikeInteraction` is present on any interaction, the page will stop scrolling in mobile mode, and the interaction takes precendence. Especially, this is done to prevent the tooltip on how to pan the map on mobile devices to appear. This flag is documented at the end of the README.md.
 - Fix: Resolve issue of plugins being placed in either `NineLayoutTag.MIDDLE_LEFT` or `NineLayoutTag.MIDDLE_RIGHT` not being centered on the map-div.